Switch from C++20 to C++17, for greater compatibility with standard libraries
Unfortunately Clang's libc++ doesn't fully support C++20 yet.
<jthread> and <stop_token> are missing, making the code unbuildable on macOS.

As <span> is the only other C++20 header we include, let's switch to C++17.
This also allows us to sync the C++ version requirement with Mumble.

The only downside is that we bring in "boost::span" and "boost::thread".
They are lightweight dependencies though, the first is even header-only.