Get rid of Session, expose Connection and use a thread pool
When it comes to scaling, multi-client TCP servers are not trivial to optimize.

Common implementations are either:

- Single-threaded.
- Multi-threaded, with a thread spawned for each connection.

Murmur (Mumble's official server implementation) is currently the first type.
It obviously doesn't perform well with hundreds of clients, especially on SBCs.

Going with the second type when designing this library was a no-brainer.
However, it's not ideal as we strictly rely on the OS scheduler's behavior.

Robert proposed to use a thread pool so that we don't flood the OS with threads.
The idea took quite a while to implement, but the result is definitely worth it.

This commit packs a lot of changes, most notably:

- Polling moved from Socket to Monitor. Now supports epoll and wepoll (Windows).
- Session is gone, as it was pretty much just a wrapper for Connection.
- Connection is exposed: no thread, no need for destructor shenanigans anymore.
- The socket is now provided to the API user, who can then create a Connection.
- The API user can either drive the connection themselves or add it to Peer.
- The connection file descriptor is provided by the API, for maximum control.