Use GSL's "span" implementation instead of Boost's
Boost's "span.hpp" was added in b4a9a100a164c1f2259d3193f5238b4c67db8ac5.
The idea was to make sure it would be available (introduced in Boost 1.78).

Now that ~9 months have passed, boost::span should be widespread in repositories.
Unfortunately Debian (and in turn Ubuntu) is stuck with 1.74 for some reason.

Thus, let's just use GSL, which is intended to be used also as a submodule.

Co-authored-by: Robert Adam <dev@robert-adam.de>