Endian: Make header-only
This allows the compiler to fully optimize the call(s) and result(s).
When the library is compiled as big endian toNetwork() and toHost() are no-op.

Hiding the implementation didn't make sense, as it's only made of intrinsics.

Unfortunately the "_byteswap_*" intrinsics are not compatible with "constexpr":

error C3615: constexpr function 'mumble::Endian::swap' cannot result in a constant expression
note: failure was caused by call of undefined function or one not declared 'constexpr'