Include dep tests
If we are building tests for libmumble, tests for self-compiled
libraries that libmumble depends on should also be built with tests to
ensure that they work properly.

Additionally, this commit now sets the corresponding variable as a cache
entry, which also silences a cmake warning about mixing options and
non-cache variables.