Add fallback FindOpus file
Instead of doing the fallback handling inline in our CMakeLists file, we
now ship a proper FindOpus.cmake file that will be used by find_package.
We tell cmake to prefer CONFIG mode though, so our module should only be
used if that failed. Then we simply fall back to the already implemented
workaround of using pkg-config to search for Opus.

Doing it this way finally gets rid of the weird cmake warning about Opus
not being found and that a FindOpus module should be provided, only to
then go ahead and report that it found Opus (which is confusing).

Fixes #1